Parsing padded data after user data content
The archive header offset value can be used along with `file.seek()` in
the original python data, however since here we are streaming bytes, we
need to calculate the current position, parse the content and then
proceed with the rest of the padded data. This in order to be positioned
in the next MPQ\x1a magic tag for the Archive Header section.

Signed-off-by: Seb Ospina <kraige@gmail.com>